+++ lib/src/pages/pages.dart
@@ -2,3 +2,4 @@ export 'supauth_wrapper.dart';
 export 'forgot_password_page.dart';
 export 'login_page.dart';
 export 'sign_up_page.dart';
+export 'update_password_page.dart';