+++ lib/src/widgets/forms/forms.dart
@@ -1,3 +1,4 @@
 export 'login_form.dart';
 export 'sign_up_form.dart';
-export 'forgot_password.dart';
+export 'forgot_password_form.dart';
+export 'update_password_form.dart';